[h2]The Symmetry Update[/h2]
In this update, we add a mirror system for the ConstructorTool.
Even though this game is meant to be immersive and not feel like an editor, players felt like a mirror tool was missing and would make a big difference on their willingness to build large vehicles.
In an effort to keep the original immersive philosophy of the game as well as keeping the code clean and stable, the symmetry is limited to a single axis at a time, at a fixed position at the center of the first frame, and only applies to the ConstructorTool which means only for Frames and Blocks.
Nevertheless, we have decided to keep this feature available also in the Adventure mode, because we think that this is still more immersive than forcing the player to go into Creative mode and coming back to Adventure with a Blueprint.
Also, it's worth noting that the symmetry enforces a odd number of frames and even number of blocks.

Many thanks to EndoSkull for convincing us to implement the mirror system, beta testing it, and making this beautiful boat as an example!
Subscribe to his YouTube channel: https://www.youtube.com/@EndoSkull

[h2]Full Changelog:[/h2]

[list]
[*] Added a mirror system (Symmetry axis) for the Constructor Tool
[*] Added Output data to Wheel and SmallWheel (Rotation Speed, Ground Friction)
[*] Labels now send their text as output when clicking on them
[*] Enabled Admin Delete button for Owner, and no need to check 'Admin Priviledges'
[*] Bugfix duplicate frames sometimes preventing the removal of beams
[*] Bugfix button_triangle sticky click on Computer screen
[*] Bugfix /give N
[*] Bugfix BlueprintTool tone keeps playing after out of range
[/list]